Separate Faucet actions
